BAM-143 change to manual release with choices (#53)
**What has changed**

Previously, the release pipeline was triggered automatically whenever a
protocol change was merged. However, this led to situations where the
backend implementation was not yet complete.

To address this, it was suggested to stop the automatic release of the
SDK. The following changes have been implemented in this PR:

- Modified the pipeline to be triggered manually.
- Uncommented the release process for the Kotlin SDK.
- Added an option to release specific SDK languages or all SDKs
simultaneously.