BAM-164: Modify ecom spec (#57)
**What are changed**

- Removed the payment_data_json from getPaymentResponse
- Using new payment method enum
- Restructured PaymentDetailsResponse to have PaymentData and SaleData
with extra fields.
- Move PaymentStatus to be shared.

----

After discussions during the R&P meeting, the team decided to adopt a
more structured approach. As a result, additional changes have been made
to the previous version to align with this direction.

- Use the same PaymentDetails in the GetPaymentsResponse and
PaymentDetailsResponse.
- In PaymentDetails, deprecate payment_data_json, use structured models,
eg. PaymentData, SaleData.
- Deprecated some of the root level fields added in the inner level.
- Fixed the indexes to be backwards compatible. 
- Added support for the capture options

Other minor changes

- renamed amount to amount_minor_units
- moved payment_method_variant to the root of PaymentData as both card
and wallet payments have the field.
- In payment wallet, added two fields card_last_4_digits and
payment_link_id, card_last_4_digits is optional as they may missing from
the transaction.


In the last meeting, we decided to keep the name of order_metadata and
leave it freeform for our client, so that they can use it as they want.

---------

Co-authored-by: ioannis-kody <ioannis.petridis@kody.com>