Task/bam 158 define ordering spec (#58)
**What does this change do**

Initial spec for the integration ordering for comment.


Questions:

- Order Flattening: Should combos be flattened into items in the Order
message?
- Order Status Enum: There is a potential duplication in the OrderStatus
enum with both COMPLETED and CLOSED statuses. Should these be
differentiated or merged? Also, it seems there is not status in the io
OrderStatus that can be used to map COMPLETE or CLOSED. Do we need
extend the OrderStatus in IO or some other places?
- For pagination, which timestamp should I used to order the records?
date_created/date_paid/date_printed.