BAM-241: make the page size and token optional (#73)
**JIRA**
BAM-241: https://kodypay.atlassian.net/browse/BAM-241

**What does the change do?**

Fix pageSize default to 0 as it is not optional in spec. 
We would like to provide a big default value from the Ordering service
for the initial release.

page token should be optional, also itemId should optional as internal
db schema.

Add itemId to combo and addon. 

Add picked status, given current orders are auto complete. We may want
to use the picked to indicate the order is picked up by the customer in
the food square.

**Does it break anything?**

No, as no body using the endpoints yet.