BAM-512: distinguish pre auth id and payment token (#111)
## **Associated JIRA tasks**

BAM-512: https://kodypay.atlassian.net/browse/BAM-512

## **What the change does.**

Previously, we misunderstood the transToken in the Opera pre-auth
transaction, assuming it was a reference to the pre-auth itself. After
revisiting the spec, we realised it is actually the payment token
generated during the pre-auth operation.

This PR updates the responses of pre-auth operations to clearly
differentiate between the pre-auth reference and the payment token.


## **Does this change break backwards compatibility?.**

No, not implemented yet